forge fmt

+++ src/MorphoFlashloan.sol
@@ -1,29 +1,23 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.24;
+
 import {IMorphoFlashLoanCallback} from "./interfaces/morpho/IMorphoFlashLoanCallback.sol";
 import {IMorpho} from "./interfaces/morpho/IMorpho.sol";
 import {TokenTransferHelper} from "./utils/TokenTransferHelper.sol";
 
 contract MorphoFlashloan is TokenTransferHelper, IMorphoFlashLoanCallback {
     IMorpho private immutable MORPHO;
-    uint public luckyNumber;
+    uint256 public luckyNumber;
 
     constructor(address newMorpho) {
         MORPHO = IMorpho(newMorpho);
     }
 
-    function takeFlashLoan(
-        address token,
-        uint256 assets,
-        bytes calldata data
-    ) external {
+    function takeFlashLoan(address token, uint256 assets, bytes calldata data) external {
         MORPHO.flashLoan(token, assets, data);
     }
 
-    function onMorphoFlashLoan(
-        uint256 assets,
-        bytes calldata data
-    ) external override {
+    function onMorphoFlashLoan(uint256 assets, bytes calldata data) external override {
         require(msg.sender == address(MORPHO));
         address token = abi.decode(data, (address));
         _doSomethingWithFlashLoan(data);
